Add KVData
This commit adds the KVData message type.  The KVData message type
contains an arbitrary sequence of bytes used to represent a key which
then points to zero or more arbitrary strings of value.  The values
optionally contain an ordering index.  If no ordering index is
provided then the order the strings are returned to a client in is
stable.