Upgrade to gRPC and Protocol Buffers v2 API.
This is an upgrade to the supported version of protobuf for go.  The
following option is required due to an astonishingly poorly thought
through change from the gRPC team:

```
--go-grpc_opt=require_unimplemented_servers=false
```

This flag is required to ensure that compile time interface checking
continues to work, which is one of the entire points that this service
is written in Go.  This is likely the last major update to the gRPC
side of things, and its very likely that NetAuth will begin to review
JSON-RPC as a potential replacement.